Remove docs (with note) about basic/custom form support. (#6290)
### Description of the change

We recently had [a query (on
slack)](https://kubernetes.slack.com/archives/C9D3TSUG4/p1686168329696059?thread_ts=1685516346.516889&cid=C9D3TSUG4)
about the basic form support, since the person had read about it in the
docs. This PR just removes the docs about basic and custom form support
with a note that we plan to be able to bring it back.

### Benefits

Less confusion when people read the docs.