Switch CheckNamespaceExist to return a connect error as expected by client (#6319)
### Description of the change

When we switched to connect-grpc, we continued sending standard gRPC
error codes from Kubeapps-APIs, and without more information, the
[connect-gRPC machinery was converting these to code
unknown](https://connect.build/docs/go/errors/#working-with-errors).

When we check that a user has authenticated successfully by getting the
default namespace after the login dance, either an OK or a forbidden
means the request was authenticated, unauthenticated means it was not.
But instead the error was coming back with an unknown code and so
Kubeapps was not recognising this as authenticated. This was affecting
logins for users who do not have permission to get the default namespace
(see #6269 for the workaround).

I reproduced this locally (though with token auth) and so before this
change, the gRPC request to CheckNamespaceExists came bace with what
*looks* like a permission denied response - but the [grpc-status code is
2 - which is
unknown](https://grpc.github.io/grpc/core/md_doc_statuscodes.html):

```
grpc-message | rpc  error: code = PermissionDenied desc = Forbidden to get the Namespace  'default' due to 'namespaces "default" is forbidden: User  "system:serviceaccount:kubeapps:kubeapps-view" cannot get resource  "namespaces" in API group "" in the namespace "default"'
grpc-status | 2 <----- this is "Unknown"
```

After this change, the error arrives as a PermissionDenied as expected:
```
grpc-message | Forbidden  to get the Namespace 'default' due to 'namespaces "default" is  forbidden: User "system:serviceaccount:kubeapps:kubeapps-view" cannot  get resource "namespaces" in API group "" in the namespace "default"'
grpc-status | 7 <----- this is the "PermissionDenied" code
```
which Kubeapps happily recognises and assumes the user is now
authenticated.

### Benefits

People can login without requiring RBAC to get the default namespace.

### Possible drawbacks
None

### Applicable issues

<!-- Enter any applicable Issues here (You can reference an issue using
#) -->

- fixes #6269 

### Additional information

Although I was glad to find the fix here, I'm still uncertain why our CI
integration tests did not pick this up, given that the intention is that
we check login with an unprivileged user. I need to investigate that
further.

Also, this is a second attempt at fixing this after #6312 ended up
touching too many call-sites. I'll then followup switching other
call-sites and removing the old code.

Signed-off-by: Michael Nelson <minelson@vmware.com>