Update copyright year (#6531)
### Description of the change

This PR is mainly to update the copyright year to the current one just
in those files that were updated in 2023. Also, for automating this
burden, I have also added a script ([file
here](https://github.com/antgamdia/kubeapps/blob/d8cc3c8a6a003368fb00724f048339d71bbc7886/script/replace-with-current-copyright-year.sh))
with the automation. It could be run periodically from a GHA, but this
is not handled in this PR.

### Benefits

The copyright header will hold the proper year in those files we forgot
to manually update them.

### Possible drawbacks

N/A

### Applicable issues

N/A

### Additional information

Good replacement examples:

```diff
-# Copyright 2022 the Kubeapps contributors.
+# Copyright 2022-2023 the Kubeapps contributors.

# Copyright 2023 the Kubeapps contributors.

-# Copyright 2020-2022 the Kubeapps contributors.
+# Copyright 2020-2023 the Kubeapps contributors.

# Copyright 2020-2023 the Kubeapps contributors.
```

Bad replacement examples, but those situations shouldn't happen:

```diff
-# Copyright 2024 the Kubeapps contributors.
+# Copyright 2024-2023 the Kubeapps contributors.

-# Copyright 2023-2024 the Kubeapps contributors.
+# Copyright 2023 the Kubeapps contributors.

-# Copyright 2024-2025 the Kubeapps contributors.
+# Copyright 2024-2023 the Kubeapps contributors.
```

---------

Signed-off-by: Antonio Gamez Diaz <agamez@vmware.com>