Update yarn deps (#6532)
### Description of the change

This PR bumps some yarn dependencies up. However, plenty of them were
already up-to-date thanks to the `^` operator.
I don't know why Dependabot is ignoring them... perhaps we closed a PR
in the past?

### Benefits

Up-to-date deps :P

### Possible drawbacks

N/A - CI is passing now after a few tweaks in the monaco-related tests.
~Edit: I'm having some issues with the e2e tests, just some minor
changes in the selector, but I'm on it already. Hope to fix the pipeline
soon.~

### Applicable issues

N/A

### Additional information

The remaining ones are the following, but they require extra work in
order to get them to work.

#### Related to React Router v6

```
    "history": "^5.3.0",
    "react-router-dom": "^6.14.2",
```

#### Related to React 18

```
    "react": "^18.2.0",
    "@testing-library/react": "^14.0.0",
    "@types/react": "^18.2.17",
    "@types/react-dom": "^18.2.7",
    "react-dom": "^18.2.0",
    "react-test-renderer": "^18.2.0",
```

#### Others
```
    "react-redux": "^8.1.1",
```

---------

Signed-off-by: Antonio Gamez Diaz <agamez@vmware.com>