Sort packages by version instead of relying on `ChartVersions[0]` (#6588)
### Description of the change

As described in #6584, we were returning wrong versions in some corner
cases (mainly using pre-releases). It seems we were relying on
`ChartVersions[0]` with the explicit assumption that it would always
contain the latest version. However, this is not the case. Since the DB
does not know about semver (unless we used some
[extensions](https://github.com/theory/pg-semver)), the ORDER BY query
might be wrong.

This PR is replacing the occurrences of `ChartVersions[0]` in favor of a
previous semver sorting. The affected operations are:
`GetAvailablePackageDetail`, `GetAvailablePackageSummaries`,
`GetInstalledPackageSummaries` and `GetInstalledPackageDetail`.

### Benefits

The `latestVersion` returned by the API (and therefore used by the UI
-as is-, we don't sort them in client-side) will be the one that has to
be.

### Possible drawbacks

1. In this PR I have just tacked the Helm plugin, not the Flux one.
2. In the code we have relied on this assumption, not 100% of the side
effects, if any.

### Applicable issues

- fixes #6584

### Additional information

Example of a new version being properly detected:


![image](https://github.com/vmware-tanzu/kubeapps/assets/11535726/f46dd30b-fc1e-471a-88b1-c434b91ce1d5)

---------

Signed-off-by: Antonio Gamez Diaz <agamez@vmware.com>