Avoid endless spinner when no readme is provided (#6590)
### Description of the change

When packages don't have any readme file, currently the spinner is never
hidden. It gives the impression of something still unloaded, so the user
should remain in that page because sth else will be displayed soon.
However, this is not true: once `! isFetching`, if there is no readme...
it will never be available.

This PR passes the `isFetching` prop downstream so that the readme
renderer can decide what to display.

### Benefits

No more endless spinner when no readme is provided.

### Possible drawbacks

N/A


### Applicable issues

N/A

### Additional information


![image](https://github.com/vmware-tanzu/kubeapps/assets/11535726/37eb207f-84aa-49c4-97e9-4605d5f278d6)

Signed-off-by: Antonio Gamez Diaz <agamez@vmware.com>