Remove old js components: Card, Table (#6560)
## Description of the change

This PR is following up #6554 (and requires it to get merged first). In
this final one, the focus is on:

- Transform the `Card` and its sub-components to tsx.
- Transform the `Table` to tsx
- Remove the remaining test snapshots 

### Benefits

No cool before/after pics this time :P Here we have just migrated
js-based code to typescript.
However, after this PR we will able to say we finally got rid of the
legacy js-components!!

### Possible drawbacks

N/A - hopefully.

### Applicable issues

Related to https://github.com/vmware-tanzu/kubeapps/issues/6563

### Additional information

While transforming the `Card` to tsx I also had a look at the `CdsCard`
from Clarity. Unfortunately, there is no drop-in replacement; a
migration would require plenty of style-related changes. Besides, the
"native" style differs a bit from the general style we have in Kubeapps.
Below are some examples:

<details><summary>Example from the Clarity website</summary>


![image](https://github.com/vmware-tanzu/kubeapps/assets/11535726/c0b783b0-25dd-48ba-b673-abd85fddd830)

</details>

<details><summary>Quick/dirty replacement, it does not look
good</summary>


![image](https://github.com/vmware-tanzu/kubeapps/assets/11535726/34a6c2c5-d5c0-418e-83d3-6143766d4170)

</details>

<details><summary>Our current custom card</summary>


![image](https://github.com/vmware-tanzu/kubeapps/assets/11535726/4bf0def2-3a8f-4680-bf01-56f3ea390d6a)

</details>

---------

Signed-off-by: Antonio Gamez Diaz <agamez@vmware.com>