Use the OCI Catalog when syncing charts. (#6692)
### Description of the change

This PR updates the sync code to also try the OCI Catalog service for
app repositories where the repos aren't listed (after trying the VAC
index).

### Benefits

Finishes the work of #6263, although, as we've discussed, to benefit
from this (ie. be able to add the public bitnami OCI repo), we need to
add a couple more features, since currently it uses the existing OCI
Distribution API which *requires* authentication. We'd like to ensure
people don't need authentication to use a public OCI namespace, such as
the bitnami catalog.

### Possible drawbacks


### Applicable issues

<!-- Enter any applicable Issues here (You can reference an issue using
#) -->

- fixes #6263

### Additional information

<!-- If there's anything else that's important and relevant to your pull
request, mention that information here.-->

Signed-off-by: Michael Nelson <minelson@vmware.com>