Sync each app when fetched rather than waiting. (#6825)
### Description of the change

Follows on from #6804, this PR updates so that the actual Syncing
happens per chart+versions, rather than at the end of the process. Much
of the change is refactoring functions that worked with slices of charts
to work with a chart instead. The complexity is the actual refactoring
of the sync, requiring channels to iterate on the results etc.

### Benefits

We'll be able no have results in the UX much more quickly (important
with OCI imports, which take longer without a big index.yaml to simply
parse).

### Possible drawbacks

Currently this will not work as expected, as two more changes are
required:

1. Need to *update* the existing chart versions for a chart, rather than
re-writing them (with only the chart versions that were synced, as it
will do now)
2. Need to collect the charts to be deleted and pass those explicitly,
rather than deleting all charts that are not in the current set being
synced (which it currently does - as this worked when syncing
everything).

### Applicable issues

- ref #6706 

### Additional information

Last PR to follow.

---------

Signed-off-by: Michael Nelson <minelson@vmware.com>