Disable SRP action from Kubeapps (#7064)
### Description of the change

Due announcement that starting on Nov. 13 (Monday), calls to the 2.5
APIs will return a 400 error message. This means that the SRPCLI
commands to submit provenance 2.5 will begin to fail, which in turn may
fail your build.
The implementation of calls to v3 directly from Bitnami makes
unnecessary to double-check in Kubeapps development pipelines.

### Benefits

SRP process is centralized in the official assets for Kubeapps and not
necessary in the development ones.

### Possible drawbacks

N/A