Fix integration tests when running all tests (#7938)
### Description of the change

After the last refactor of the integration tests, there is a bug when
running the tests passing `all` as the `TESTS_GROUP` variable. The
script is trying to run the tests under `all` directory for every test
group instead of running the tests in each corresponding directory. This
PR fixes that.

### Benefits

Intregration tests works properly when `all` is passed as the
`TESTS_GROUP` to run.

### Possible drawbacks

<!-- Describe any known limitations with your change -->

### Applicable issues

N/A

### Additional information

N/A

Signed-off-by: Jesús Benito Calzada <jesus.benito@broadcom.com>