Add deprecation notice to indicate project will no longer be maintained (#8374)
After many years of growth and milestones, the time has come for the
next chapter of Kubeapps. The current maintainers are concentrating
their efforts on new projects. We are looking for volunteers to help us
maintain the project going forward. Without new volunteers, the project
will be deprecated and the repo archived on August 8th, 2025.

Signed-off-by: Carlos Rodríguez Hernández <carlos.rodriguez-hernandez@broadcom.com>